refactor code, phase 1 finished
