--- downstream/TextSGC/requirements.txt
@@ -1,1 +1,0 @@
-hyperopt==0.1.1
