--- downstream/TextSGC/tuned_result/20ng.SGC.tuning.txt
@@ -1,1 +1,0 @@
-0.0013576034109430972
