--- downstream/TextSGC/tuned_result/R52.SGC.tuning.txt
@@ -1,1 +1,0 @@
-3.39184266146822e-06
