--- downstream/TextSGC/tuned_result/R8.SGC.tuning.txt
@@ -1,1 +1,0 @@
-0.000555422439238392
