--- downstream/TextSGC/tuned_result/mr.SGC.tuning.txt
@@ -1,1 +1,0 @@
-0.0031331167329072614
