--- downstream/TextSGC/tuned_result/ohsumed.SGC.tuning.txt
@@ -1,1 +1,0 @@
-0.003911359029974794
