+++ egs/data/ind.citeseer.allx
