+++ egs/data/ind.citeseer.ally
