+++ egs/data/ind.citeseer.graph
