+++ egs/data/ind.citeseer.tx
