+++ egs/data/ind.citeseer.ty
