+++ egs/data/ind.citeseer.x
