+++ egs/data/ind.citeseer.y
