+++ egs/data/ind.cora.allx
