+++ egs/data/ind.cora.ally
