+++ egs/data/ind.cora.tx
