+++ egs/data/ind.cora.ty
