+++ egs/data/ind.cora.x
