+++ egs/data/ind.cora.y
