+++ egs/data/ind.pubmed.ally
