+++ egs/data/ind.pubmed.tx
