+++ egs/data/ind.pubmed.ty
