+++ egs/sgc-tuned/citeseer.txt
