+++ egs/sgc-tuned/cora.txt
