+++ requirements.txt
@@ -1,6 +1,7 @@
 numpy
 scipy
-networkx
+networkx==1.11
+hyperopt==0.1.1
 scikit-learn
 torch
 pytorch_lightning
